multimodal-vqa-system/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0
streamlit>=1.25.0
pillow>=9.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf<4.0.0
